pyproject.toml
[tool.poetry]
name = "sentinel"
version = "0.1.0"
description = "Autonomous agent orchestration system for Jira ticket implementation"
authors = ["Sentinel Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
pyyaml = "^6.0.1"
requests = "^2.31.0"
langchain = "^0.1.0"
langgraph = "^0.0.20"
pydantic = "^2.5.0"
claude-agent-sdk = "^0.1.20"
python-dotenv = "^1.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
ruff = "^0.1.9"
mypy = "^1.7.1"
pytest-asyncio = "^0.23.2"
types-pyyaml = "^6.0.12.20250915"
types-requests = "^2.32.4.20260107"
pytest-cov = "^7.0.0"

[tool.poetry.scripts]
sentinel = "src.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
